Moved the metrics listener out of its own package.
